+++ src/spoint.cpp
@@ -38,7 +38,7 @@ void spoint::operator/=(short b)
   y /= b;
 }
 
-float spoint::length(void) const { return (sqrt(x * x + y * y)); }
+float spoint::length(void) const { return ((float) sqrt(x * x + y * y)); }
 
 void spoint::unit(void)
 {
